+++ CHANGELOG.md
@@ -7,12 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - Searching within involvement admin will now search more fields.
+- A new form field type for person numbers
 ### Changed
 - Switch to the materialize CSS/JS framework.
 ### Removed
 - Old migrations in the website app.
 ### Fixed
 - Searching for positions within the admin interface will no longer cause an error.
+- Allow person numbers to be "T-numbers"
 
 ## [0.2.0] - 2017-04-07
 ### Added